feat: move the operation annotations to repositoryl
